requirements.txt
azure-ai-documentintelligence==1.0.1
azure-core==1.30.0
azure-search-documents==11.6.0b4
python-dotenv==1.0.0
langchain==0.2.1
langchain-text-splitters==0.2.4
openai==1.30.5
tiktoken==0.7.0
pandas==2.2.2
Flask==3.1.0
azure-ai-evaluation
openpyxl